feat:H9 DEBUG

+++ src/components/Students/Student.tsx
@@ -10,18 +10,23 @@ interface Props {
 
 const Student = ({ student, onDelete }: Props): React.ReactElement => {
   const onDeleteHandler = (): void => {
-    onDelete(student.id);
+    if (confirm('Удалить студента?')) {
+      debugger;
+      console.log('onDeleteHandler', student.id);
+      onDelete(student.id);
+    }
+    
   };
 
   return (
     <div className={`${styles.Student} ${student.isDeleted ? styles['--isDeleted'] : '' }`}>
       {student.id}
       {' - '}
-      {student.last_name}
+      {student.lastName}
       {' '}
-      {student.first_name}
+      {student.firstName}
       {' '}
-      {student.middle_name}
+      {student.middleName}
       {' '}
       <button onClick={onDeleteHandler}>Удалить</button>
     </div>
